frontend/requirements.txt
# Memory Extraction & Personality Engine - Streamlit Frontend

streamlit
requests
python-dotenv